+++ packages/components/choropleth/src/index.js
@@ -1,3 +1,3 @@
 export {default as ChoroplethG} from './ChoroplethG.svelte';
 export {default as ChoroplethDiv} from './ChoroplethDiv.svelte';
-export * from './shared';
+export * from './shared.js';